# backend/server.js
const http = require('http');
const express = require('express');
const mongoose = require('mongoose');
const cors = require('cors');
const dotenv = require('dotenv');
const { Server } = require('socket.io');
const TrafficEvent = require('./src/models/TrafficEvent');
const CameraDetection = require('./src/models/CameraDetection');
const startMockTrafficStream = require('./src/mockDataGenerator');
const startMqttBridge = require('./src/mqttBridge');
const { startSimulators } = require('./src/sensorSimulators');
const { initPostgres } = require('./src/postgresClient');
const { router: deviceRouter, verifyDeviceToken } = require('./src/auth');

dotenv.config();
const app = express();
app.use(cors());
app.use(express.json());

app.use((err, req, res, next) => {
  if (err instanceof SyntaxError && err.status === 400 && 'body' in err) {
    return res.status(400).json({ error: 'invalid_json', message: err.message });
  }
  next(err);
});

const PORT = process.env.PORT || 4000;
const MONGO_URI = process.env.MONGO_URI || 'mongodb://mongo:27017/trafficdb';

app.get('/api/health', (req, res) => {
  res.json({ status: 'ok', uptime: process.uptime() });
});

app.get('/api/detections', async (req, res) => {
  const detections = await CameraDetection.find().sort({ timestamp: -1 }).limit(50);
  res.json(detections);
});

app.get('/api/signals', (req, res) => {
  if (mqttClient && typeof mqttClient.getSignalStates === 'function') {
    return res.json(mqttClient.getSignalStates());
  }
  res.json({});
});

app.post('/api/signals', verifyDeviceToken, (req, res) => {
  const { intersection = 'A1', phase = 'green', duration = 30 } = req.body || {};
  if (!mqttClient || typeof mqttClient.publishSignalCommand !== 'function') {
    return res.status(503).json({ error: 'MQTT bridge unavailable' });
  }
  mqttClient.publishSignalCommand(intersection, phase, duration);
  res.json({ intersection, phase, duration });
});

app.get('/api/traffic', async (req, res) => {
  const events = await TrafficEvent.find().sort({ timestamp: -1 }).limit(100);
  res.json(events);
});

app.get('/api/traffic/summary', async (req, res) => {
  const summary = await TrafficEvent.aggregate([
    { $sort: { timestamp: -1 } },
    { $limit: 50 },
    {
      $group: {
        _id: null,
        averageSpeed: { $avg: '$averageSpeed' },
        averagePollution: { $avg: '$pollutionIndex' },
        totalVehicles: { $sum: '$vehicleCount' }
      }
    }
  ]);

  res.json(summary[0] || { averageSpeed: 0, averagePollution: 0, totalVehicles: 0 });
});

app.post('/api/traffic', verifyDeviceToken, async (req, res) => {
  const payload = req.body;
  const event = new TrafficEvent({
    intersection: payload.intersection || 'intersection-1',
    vehicleCount: payload.vehicleCount || 0,
    averageSpeed: payload.averageSpeed || 0,
    pollutionIndex: payload.pollutionIndex || 0,
    signalPhase: payload.signalPhase || 'green',
    congestionLevel: payload.congestionLevel || 'moderate',
    sensorType: payload.sensorType || 'aggregated',
    source: 'api'
  });

  await event.save();
  io.emit('trafficUpdate', event);
  res.status(201).json(event);
});

app.post('/api/predict', async (req, res) => {
  const inferenceUrl = process.env.INFERENCE_URL || 'http://inference:5001';
  try {
    const response = await fetch(`${inferenceUrl}/predict`, {
      method: 'POST',
      headers: { 'Content-Type': 'application/json' },
      body: JSON.stringify(req.body)
    });
    const data = await response.json();
    return res.status(response.status).json(data);
  } catch (error) {
    console.error('Predict proxy error:', error.message || error);
    return res.status(500).json({ error: 'Prediction service unavailable' });
  }
});

// device registration and auth
app.use('/api/devices', deviceRouter);

const server = http.createServer(app);
const io = new Server(server, {
  cors: {
    origin: '*',
    methods: ['GET', 'POST']
  }
});

// Prometheus metrics
const client = require('prom-client');
const collectDefaultMetrics = client.collectDefaultMetrics;
collectDefaultMetrics();
const httpRequestCounter = new client.Counter({ name: 'http_requests_total', help: 'Total HTTP requests' });

app.use((req, res, next) => {
  httpRequestCounter.inc();
  next();
});

app.get('/metrics', async (req, res) => {
  res.set('Content-Type', client.register.contentType);
  res.end(await client.register.metrics());
});

io.on('connection', (socket) => {
  console.log('Client connected:', socket.id);
  socket.on('disconnect', () => {
    console.log('Client disconnected:', socket.id);
  });
});

let mqttClient = null;

async function start() {
  await mongoose.connect(MONGO_URI, {
    useNewUrlParser: true,
    useUnifiedTopology: true
  });

  server.listen(PORT, () => {
    console.log(`Backend running on http://localhost:${PORT}`);
  });

  // start existing mock stream saving to MongoDB + Socket.IO
  startMockTrafficStream(io, TrafficEvent);

  // start MQTT bridge to accept sensor data via MQTT and publish to Socket.IO
  try {
    mqttClient = startMqttBridge(io, TrafficEvent);
  } catch (err) {
    console.warn('MQTT bridge failed to start', err.message || err);
  }

  // run sensor simulators when enabled (Docker Compose sets SIMULATE_SENSORS=1)
  if (process.env.SIMULATE_SENSORS === '1') {
    startSimulators();
  }

  // initialize Postgres connection (optional)
  try {
    const { sequelize } = initPostgres();
    sequelize.authenticate().then(() => console.log('Postgres connected')).catch(() => {});
  } catch (err) {
    // ignore if not configured
  }
}

start().catch((error) => {
  console.error('Startup error:', error);
  process.exit(1);
});