# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Traffic Monitoring API
  version: 1.0.0
  description: REST API for AI-powered traffic monitoring, forecasting, and signal control.
paths:
...
  /api/traffic:
    get:
      summary: List recent traffic events
      responses:
        '200':
          description: Traffic events
    post:
      summary: Ingest traffic event from authenticated device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                intersection:
                  type: string
                vehicleCount:
                  type: number
                averageSpeed:
                  type: number
                pollutionIndex:
                  type: number
                signalPhase:
                  type: string
                congestionLevel:
                  type: string
                sensorType:
                  type: string
      responses:
        '201':
          description: Created
...
  /api/signals:
    get:
      summary: Current signal phase states
      responses:
        '200':
          description: Signal states by intersection
    post:
      summary: Publish a signal phase command
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                intersection:
                  type: string
                phase:
                  type: string
                duration:
                  type: integer
      responses:
        '200':
          description: Command accepted
  /api/predict:
    post:
      summary: Get short-term LSTM forecast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                window:
                  type: array
                  items:
                    type: number
      responses:
        '200':
          description: Prediction result
        '400':
          description: Bad request
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer